requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
docker==7.1.0
idna==3.10
numpy==2.0.2
python-nmap==0.7.1
requests==2.32.3
scapy==2.6.1
urllib3==2.3.0
